updated postman documentation

+++ controllers/blog/deleteBlog.js
@@ -1,13 +1,21 @@
 const BlogArticle = require("../../models/BlogArticle");
 const BlogContent = require("../../models/BlogContent");
+const mongoose = require("mongoose");
 
 const deleteBlog = async (req, res) => {
   try {
-    const blog = await BlogArticle.findByIdAndDelete(req.params.id);
-
-    if (!blog) {
+    const writtenBy = req.user._id;
+    const toDeleteBlog = await BlogArticle.findById(req.params.id);
+    if (!toDeleteBlog) {
       return res.status(404).json({ message: "Blog not found" });
     }
+    if (
+      writtenBy != toDeleteBlog.writtenBy.toString() &&
+      req.user.isAdmin === false
+    ) {
+      return res.status(404).json({ message: "unauthorized" });
+    }
+    const blog = await BlogArticle.findByIdAndDelete(req.params.id);
 
     await BlogContent.deleteMany({
       _id: { $in: blog.contents.map((c) => c.id) },
